refactor: separate logic
