+++ p2p_wallet/Scenes/Main/ReceiveToken/NetworkSelection/ConfirmReceivingBitcoin/ConfirmReceivingBitcoin.ViewController+CreateRenBTC.swift
@@ -1,34 +1,13 @@
 //
-//  ConfirmReceivingBitcoin.ViewController+Extensions.swift
+//  ConfirmReceivingBitcoin.ViewController+CreateRenBTC.swift
 //  p2p_wallet
 //
 //  Created by Chung Tran on 22/04/2022.
 //
 
 import Foundation
-import RxCocoa
-import RxSwift
 
 extension ConfirmReceivingBitcoin.ViewController {
-    func topUpRequiredView() -> BEVStack {
-        BEVStack(spacing: 12) {
-            ReceiveToken.textBuilder(
-                text: L10n.aToReceiveBitcoinsOverTheBitcoinNetwork(L10n.renBTCAccountIsRequired)
-                    .asMarkdown()
-            )
-            ReceiveToken.textBuilder(
-                text: L10n.yourWalletListDoesNotContainARenBTCAccountAndToCreateOne(L10n
-                    .youNeedToMakeATransaction)
-                    .asMarkdown()
-            )
-            ReceiveToken.textBuilder(
-                text: L10n.youToPayForAccountCreationButIfSomeoneSendsRenBTCToYourAddressItWillBeCreatedForYou(L10n
-                    .donTHaveFunds)
-                    .asMarkdown()
-            )
-        }
-    }
-
     func createRenBTCView() -> BEVStack {
         BEVStack(spacing: 12) {
             ReceiveToken.textBuilder(
@@ -106,32 +85,7 @@ extension ConfirmReceivingBitcoin.ViewController {
         }
     }
 
-    func buttonsView() -> UIView {
-        BEVStack(spacing: 10) {
-            WLStepButton.main(image: .add, text: L10n.topUpYourAccount)
-                .setup { view in
-                    viewModel.accountStatusDriver
-                        .map { $0 != .topUpRequired }
-                        .drive(view.rx.isHidden)
-                        .disposed(by: disposeBag)
-                }
-
-            WLStepButton.sub(text: L10n.shareYourSolanaNetworkAddress)
-                .setup { view in
-                    viewModel.accountStatusDriver
-                        .map { $0 != .topUpRequired }
-                        .drive(view.rx.isHidden)
-                        .disposed(by: disposeBag)
-                }
-
-            WLStepButton.main(text: "Pay 0.509 USDC & Continue")
-                .setup { view in
-                    viewModel.accountStatusDriver
-                        .map { $0 != .payingWalletAvailable }
-                        .drive(view.rx.isHidden)
-                        .disposed(by: disposeBag)
-                }
-        }
-        .padding(.init(only: .bottom, inset: 18))
+    func createRenBTCButton() -> UIView {
+        WLStepButton.main(text: "Pay 0.509 USDC & Continue")
     }
 }
